# Improve transaction REST API query performance