--- a/hedera-mirror-importer/src/main/resources/db/migration/v1/V1.46.5__rename_consensus_ns.sql
+++ b/hedera-mirror-importer/src/main/resources/db/migration/v1/V1.46.5__rename_consensus_ns.sql
@@ -1,0 +1,3 @@
+-- Rename transaction consensus_ns column to consistently uses consensus_timestamp
+alter table transaction
+    rename column consensus_ns to consensus_timestamp;