--- a/hedera-mirror-importer/src/main/resources/db/migration/v2/V2.0.0__time_scale_init.sql
+++ b/hedera-mirror-importer/src/main/resources/db/migration/v2/V2.0.0__time_scale_init.sql
@@ -366,7 +366,7 @@ comment on table topic_message is 'Topic entity sequenced messages';
 -- transaction
 create table if not exists transaction
 (
-    consensus_ns           bigint   not null,
+    consensus_timestamp    bigint   not null,
     type                   smallint not null,
     result                 smallint not null,
     payer_account_id       bigint   not null,