--- a/hedera-mirror-importer/src/main/resources/db/migration/v2/V2.0.2__time_scale_index_init.sql
+++ b/hedera-mirror-importer/src/main/resources/db/migration/v2/V2.0.2__time_scale_index_init.sql
@@ -156,13 +156,13 @@ create unique index if not exists topic_message__topic_num_realm_num_seqnum
 
 -- transaction
 alter table if exists transaction
-    add primary key (consensus_ns);
+    add primary key (consensus_timestamp);
 create index if not exists transaction__transaction_id
     on transaction (valid_start_ns, payer_account_id);
 create index if not exists transaction__payer_account_id
     on transaction (payer_account_id);
 create index if not exists transaction_type
-    on transaction (type, consensus_ns desc);
+    on transaction (type, consensus_timestamp desc);
 
 -- transaction_signature
 create index if not exists transaction_signature__entity_id