Signed-off-by: Nana-EC <nana.essilfie-conduah@hedera.com>

**Description**:
The `transactions.js` logic utilizes heavy inner and outer queries to provide results to the `api/v1/accounts` and `api/v1/transactions` endpoints.
These queries are inefficient with multiple joins and don't lend themselves to distributed querying.

In #2273 it was possible to modify the queries to improve performance, this change adopts those improvements

- Convert transaction sub query to CTE, with support for an additional CTE when transactions need to be further filtered
- Break out transfer tables into separate CTEs to support distribution
- Utilize CTE result for base of transactions and avoid extra join on transaction table again
- Rename transaction table `consensus_ns` column to standard `consensus_timestamp`

**Related issue(s)**:

Fixes #2676 

**Notes for reviewer**:
Performance results after runs against a local db with a large data set (~6.5 million transactions at ~1.5 GB).
Summary: There's a slight improvement in response times of around 500-1000ms on average.
Slowest times are improved and the distribution of times is more narrow and smaller.
Additionally, the structure of queries support multi node distribution so we should see even more improvements  in such an environment.

### Before
1. GET http://localhost:8551/api/v1/transactions
Summary:
  Total:	218.5966 secs
  Slowest:	15.3077 secs
  Fastest:	0.7489 secs
  Average:	4.3533 secs
  Requests/sec:	45.7464


Response time histogram:
  0.749 [1]	|
  2.205 [52]	|
  3.661 [3578]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  5.117 [4997]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  6.572 [498]	|■■■■
  8.028 [211]	|■■
  9.484 [112]	|■
  10.940 [160]	|■
  12.396 [218]	|■■
  13.852 [55]	|
  15.308 [77]	|■


Latency distribution:
  10% in 2.8894 secs
  25% in 3.4509 secs
  50% in 3.8418 secs
  75% in 4.3915 secs
  90% in 5.8090 secs
  95% in 9.5823 secs
  99% in 13.5345 secs


2. GET http://localhost:8551/api/v1/transactions?account.id=0.0.9
Summary:
  Total:	177.2672 secs
  Slowest:	7.0617 secs
  Fastest:	0.6213 secs
  Average:	3.5333 secs
  Requests/sec:	56.4120


Response time histogram:
  0.621 [1]	|
  1.265 [32]	|
  1.909 [45]	|■
  2.553 [951]	|■■■■■■■■■■■
  3.197 [2685]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.841 [3564]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  4.486 [1270]	|■■■■■■■■■■■■■■
  5.130 [755]	|■■■■■■■■
  5.774 [402]	|■■■■■
  6.418 [101]	|■
  7.062 [127]	|■


Latency distribution:
  10% in 2.5489 secs
  25% in 2.8812 secs
  50% in 3.4258 secs
  75% in 3.8947 secs
  90% in 4.8643 secs
  95% in 5.2828 secs
  99% in 6.5154 secs

3. GET http://localhost:8551/api/v1/transactions?type=credit
Summary:
  Total:	160.6748 secs
  Slowest:	5.0749 secs
  Fastest:	0.4776 secs
  Average:	3.2033 secs
  Requests/sec:	62.2375


Response time histogram:
  0.478 [1]	|
  0.937 [9]	|
  1.397 [23]	|
  1.857 [34]	|
  2.317 [91]	|■
  2.776 [3111]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.236 [2869]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.696 [1559]	|■■■■■■■■■■■■■■■■■■■■
  4.155 [803]	|■■■■■■■■■■
  4.615 [1269]	|■■■■■■■■■■■■■■■■
  5.075 [168]	|■■


Latency distribution:
  10% in 2.5470 secs
  25% in 2.7086 secs
  50% in 3.0369 secs
  75% in 3.5664 secs
  90% in 4.2817 secs
  95% in 4.4389 secs
  99% in 4.7485 secs

### After
1. GET http://localhost:5551/api/v1/transactions

Summary:
  Total:	150.6810 secs
  Slowest:	4.5372 secs
  Fastest:	0.8895 secs
  Average:	3.0039 secs
  Requests/sec:	66.3654


Response time histogram:
  0.889 [1]	|
  1.254 [16]	|
  1.619 [28]	|
  1.984 [30]	|
  2.349 [245]	|■■■
  2.713 [3309]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.078 [2790]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.443 [1207]	|■■■■■■■■■■■■■■■
  3.808 [1316]	|■■■■■■■■■■■■■■■■
  4.172 [819]	|■■■■■■■■■■
  4.537 [184]	|■■


Latency distribution:
  10% in 2.4608 secs
  25% in 2.6174 secs
  50% in 2.8675 secs
  75% in 3.3733 secs
  90% in 3.8111 secs
  95% in 3.9225 secs
  99% in 4.2746 secs


2. GET http://localhost:5551/api/v1/transactions?account.id=0.0.9
Summary:
  Total:	135.4116 secs
  Slowest:	4.0858 secs
  Fastest:	0.6763 secs
  Average:	2.6969 secs
  Requests/sec:	73.8489


Response time histogram:
  0.676 [1]	|
  1.017 [20]	|
  1.358 [26]	|
  1.699 [36]	|
  2.040 [16]	|
  2.381 [730]	|■■■■■■
  2.722 [5275]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.063 [2244]	|■■■■■■■■■■■■■■■■■
  3.404 [1407]	|■■■■■■■■■■■
  3.745 [50]	|
  4.086 [134]	|■


Latency distribution:
  10% in 2.3944 secs
  25% in 2.4604 secs
  50% in 2.5707 secs
  75% in 2.9764 secs
  90% in 3.1227 secs
  95% in 3.2138 secs
  99% in 3.8527 secs


3. GET http://localhost:5551/api/v1/transactions?type=credit
Summary:
  Total:	145.4790 secs
  Slowest:	4.3652 secs
  Fastest:	0.4937 secs
  Average:	2.9012 secs
  Requests/sec:	68.7384


Response time histogram:
  0.494 [1]	|
  0.881 [6]	|
  1.268 [30]	|
  1.655 [33]	|
  2.042 [34]	|
  2.429 [617]	|■■■■■■
  2.817 [4038]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.204 [2682]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.591 [2046]	|■■■■■■■■■■■■■■■■■■■■
  3.978 [287]	|■■■
  4.365 [156]	|■■


Latency distribution:
  10% in 2.4610 secs
  25% in 2.5751 secs
  50% in 2.8454 secs
  75% in 3.2056 secs
  90% in 3.4573 secs
  95% in 3.5782 secs
  99% in 4.1577 secs

**Checklist**

- [ ] Documented (Code comments, README, etc.)
- [ ] Tested (unit, integration, etc.)